Add a symlink to run migration 0004 again as 0006.
I modified 0004 to handle the case where the change it is making
(changing repo_id's on package groups to match the repo they are
associated with) results in a DuplicateKeyError. To handle this case, we
simply remove the offending package group since there is already one
that represents what it would become.