1093745: use filename when appropriate instead of relativepath in RPM model obj
Previously, relativepath was being used when querying for the location of an
RPM locally. This caused issues when the same RPM was synced from two sources,
since the source for the symlink would be incorrect, causing an error.

Relativepath has been changed to filename, so the symlink source is now correct.