1095829 - strip repomd.xml from treeinfo when appropriate
Some treeinfo files (notably Fedora) contain sha256 sums for
repodata/repomd.xml. This causes pulp to treat repomd.xml as content instead of
metadata, which causes the newly generated repomd.xml to not be able to be
linked during publish.

For now, we strip references to 'repodata/repomd.xml' when saving treeinfo
files to avoid this. We will follow up with the Fedora team to see why this
checksum is being included.