1101168 - use metadata when computing RPM filename
Previously, an uploaded RPM's filename was used as the filename when saving
under `/var/lib/pulp/content/rpm`. This led to problems if the filename was not
the same as the standard RPM naming convention or if the upload was via a
temporary file.

This commit changes the file saving behavior for RPMs so the on-disk filename
is generated via the RPM NVRA metadata. This should not affect upgrades from
2.3 since it only touches uploading of new RPMs.

Another implementation would have been to fix clients so the uploaded filename
is correct. However, this would leave naming of RPMs up to various clients and
not to the server.