1099600 - fix treeinfo files during upgrades
If a user has a 2.3 repo with a treeinfo file that contains repomd.xml checksum
definitions, the upgrade to 2.4 will fail during publish.

To fix this, we need to fix on-disk treeinfo files to remove the offending
checksum definition, and then fix any distribution units to remove the defined
file. The actual repomd.xml files are left on-disk and get cleaned up during
the next reposync.

Note that this fix needs to happen before the "new yum distributor" migration,
which is why the migration script ordering was juggled a bit.