1125388 - ensure we save storage_path when saving units
The code inside check_all_and_associate() works with units to ensure they are
downloaded if the unit exists but the file is not on-disk, and to associate the
unit with the repo if the unit already exists.

The unit gets re-created from an RPM model object, which can create issues
since the existing storage_path on the unit may not be what pulp thinks the
storage_path should be. For example, if a user upgrades from 2.3 to 2.4, the
storage_path of new units has changed but old units stays the same.

This fix ensures that the storage_path is copied over onto the unit before
saving.