1150714 - delete old distribution units when syncing
Distribution units have a unit key that contains the OS version, as read from
the treeinfo file.

When a new OS version comes out, a directory containing the most recent OS bits
will have a new treeinfo file with a different version and hence different unit
key. However, the old unit is not removed from the repo. This means there are
two distribution units for a single repo, which causes publish to fail.

This patch checks to see if the unit keys of any existing distributions are
different than those of the unit just saved. If so, the incorrect units are
removed, leaving only the latest correct unit.