1185868 - create fewer dictionaries when depsolving
Previously, calling `copy_rpms()` with dep solving  enabled would result in a
large number of new dictionaries. For example, copying a single erratum from
RHEL 6.6 to 6.1 would result in 1.3 million new dicts. This ate a large amount
of memory.

Instead, flatten the "provides" metadata on the unit when doing dep solving
since we only need the name of the provide, not the full dict. This reduces the
number of created dicts to 81K for the scenario mentioned above, and results in
a ~70% memory savings.

Credit for this fix goes to @mhrivnak.