770: Only apply packages in errata when they are actually available
Previously, if an erratum contained packages with the same name but different
versions, Pulp would attempt to install the latest version of said package when
applying an erratum to a system. This behavior is incorrect since an erratum
might contain multiple packages across repos.

Instead, we need to filter out any packages that we do not have access to
before applying the erratum.

re #770