treat 'epoch' as '0' when it does not exist or is null
Some updateinfo.xml metadata does not contain the `epoch` field in package
records. In these cases, Pulp will not be able to find the appropriate RPM.

In these cases, we can assume that the epoch should be '0'. This allows Pulp to
find the RPM when applying packages to a system.

fixes #833