Fix consumer binding to RPM repositories.
This addresses two separate issues. The first is the the Pulp server
config was used incorrectly when creating the consumer bind payload. The
second is that it was previously impossible to find a repository if it
did not have a 'name', which is not a requirement for creating a
repository in Pulp.

closes #1426