Revert ``Errata.pushcount`` to be ``StringField``.
In commit d3b3e5c4d58b98930c7041489274a98c229227aa the pushcount field
in Errata was changed to be an IntField. Although the field *should* be
an integer, many errata do not provide the pushcount. Since this breaks
syncing with these repos, the change has been reverted until it can be
addressed properly. This effort is tracked in issue #1458.

closes #1455