RepoURLModifier is more IDE- and user-friendly
In addition to explicitly declaring the kwargs to `__init__` and
`__call__`, I took a step back and considered if there was any value
to storing `path_append` and `ensure_trailing_slash` as stateful defaults.
Based on the current usage and my best judgement, there was no value in
that, and in fact having the same signature for both methods only made
things more confusing. Hopefully the stateful nature of of this class is
more clear, so in addition to being IDE-friendly, it's much easier for
humans to grok.

https://pulp.plan.io/issues/1442
fixes #1442