iso distributor no longer fails with wrong type of Repository object
This is the remaining work outside platform to fix #1480. The conversion from
old to new style repository was apparently not complete. It was simplest and
safest to go back to passing around the old style object, especially since the
workflow passes back and forth between the parent and child classes, in
platform and this plugin respectively.

re #1480
https://pulp.plan.io/issues/1480