set uploaded unit href to unit filename in repo
When uploading a unit, the href in the generated xml's location tag is
based on the content unit file path, which is an autogenerated unique
value. This sets the href back to the unit filename, which matches the
published repo layout and allows uploaded units to again be downloaded
from the published repo.

fixes #1556
https://pulp.plan.io/issues/1556