Use minimal copies of the RPM units when associating.
This commit adjusts
pulp_rpm.plugins.importers.yum.associate._associate_unit() to cause
it to return minimal copies of the RPM units it processes instead
of the full units. This is a hack to reduce the RAM usage of the
associated_units set in the associate() call in the same module.
That set stores the complete list of RPMs (and other Unit types)
that are to be associated during a task, and the full RPM metadata
for a large repository can easily consume several GBs of RAM.

With this commit, the set still stores the full metadata of the
other units, and continues to store the full list of Units to be
copied so it still does not win the most efficient algorithm of the
year award.

https://pulp.plan.io/issues/1662

fixes #1662