Include "TS_TRUEINSTALL" as a valid package state
When installing or updating packages in a consumer, pulp inspects the
yum transaction report to decide how to report its actions in the task
result. pulp was excluding the "TS_TRUEINSTALL" state, which yum uses as
the state of a transaction that updated or installed one of the special
packages that can be installed multiple times, such as the kernel.

Including (or rather, not excluding) that state in pulp's transaction
reporting lets those packages be included in pulp's task results.

https://pulp.plan.io/issues/1138
fixes #1138