Allow for new `incremental_export_repomd` exporter distributor config option
Previously, incremental exports were in a custom format where each erratum and
RPM were written to disk, along with a set of JSON files with unit metadata.
This required special handling of incremental imports that was significantly
different a "regular" repo sync.

This patch adds a new `incremental_export_repomd` distributor option. When this
is set, the incremental repo export generates metadata so that the on-disk
files can be synced via the typical yum sync process. Note that you would not
want to remove old units when doing this, since it would remove any units not
in the incremental :koala:.

Fixes #1543
https://pulp.plan.io/issues/1543