yum_distributor now uses configured checksum type for all metadata.
A lot of model-related code was moved into models.py from other places. In
addition to being more object-oriented, it made that code accessible from
multiple places instead of being isolated somewhere, such as in the upload
code. Being able to use the code from multiple places was the primary reason
for moving the code in this PR.

re #1618
https://pulp.plan.io/issues/1618