Overwriting a metadata unit during copy disassociates first
The metadata units have an overwrite functionaity whereby newer
units overwrite older units if they have the same unit key. The
unit being overwritten is deleted and the new one is added. The
unit being deleted did not have disassociate called so the
unit count would be wrong for these unit types after multiple
copies.

This comes with a migration which regenerates the unit counts
for all repos fixing any incorrect data as a result of the bug.

HEAD of pulp-smash passes and all unit tests pass

https://pulp.plan.io/issues/1979
fixes #1979