Erratum publish reads repo nevra into memory
The Errata publish performance became a problem recently
due to lots of db calls being made as Errata metadata is filtered
to only refer to packages in this specific repo.

This PR causes all the repo's NEVRA to be stored in memory once
and all filtering to work out of that data structure. This avoids
searching the db everytime for the NEVRA specified by the erratum
so it does much less work than before.

https://pulp.plan.io/issues/1949
fixes #1949