Make sample JSON code valid & pretty
According to the JSON specification, objects and arrays must not have
trailing commas. The following are invalid:

    {"foo": "bar",}
    [1, 2, 3,]

Make the sample JSON code in the rsync distributor documentation pages
valid. Also, reformat the JSON code samples with `json_reformat`, so
that indentation is consistent, instead of using a mixture of one, three
and four spaces.