Updates API and pulp-admin to accept relative url configuration for ISO content
ISO distributor continues to use the repository id as the default relative URL for the repository. It is now
possible to specify the relative url for the ISO distributor using the API and pulp-admin iso repo create/update
commands.
When then ISO distributor is used as predistributor, the ISO rsync distributor uses the relative url of the
predistributor when publishing.

fixes #2132
https://pulp.plan.io/issues/2132