Reclassify signature verification as signature and key ID filtering
This is a cherry-pick of 409687edaa792427c7876c815d517b66f05fe25d, which
I didn't merge forward properly to master. :(

The features introduced in #1991 (https://pulp.plan.io/issues/1991) only act
as filters based on whether or not a package is signed, and the short key ID
of the key used to generate that signature. This changes any reference to
"verification" introduced with those changes to "GPG Key ID filtering", and
in general attempted to clarify that this is not a security feature, while
still leaving it possible to (hopefully) do GPG package signature verification
in a future version.

fixes #2188
https://pulp.plan.io/issues/2188