Amend the recursive associate call
The associate recursive call is broken in the content category code path;
the return value is a tuple containing the list of associated and the list
of failed units, that can't be used directly when updating the
associated_units set.

The patch corrects this by using the tuple items to extend the
associated_units and failed_units sets.  It also removes the set->list
conversion at the return statements for easier handling of the recursive
call as well as for performance benefits.

Fixes: #3787
https://pulp.plan.io/issues/3787