Added quickstart page
The quickstart page gets the user going with pulplift and helps them
check the status API. From there it links to the already great workflows
page.

This also fixed the readthedocs theme to be used which is important
because only that theme renders the left-panel in a way that you can see
this PR as it appears on the website. If you don't have the
sphinx_rtd_theme package installed you get the default theme as a
fallback.

https://github.com/rtfd/sphinx_rtd_theme

https://pulp.plan.io/issues/4620
closes #4620