Switch tasks to use pks
The copy and upload tasks were passing large objects instead of passing
pks. This switches those (and their corresponding dispatch code) to use
pks instead, and the objects are re-queried in the task. Some of the
tasks didn't need an update because they already used that form. All
tasks were checked.

https://pulp.plan.io/issues/4876
closes #4876