Replace the Pulp version of module dependency code with MBS version
Use Module Build Service code to create libsolv dependencies for
modules. The code is better commented, cleaner, possibly less buggy and
has had more eyes on it. It is otherwise almost identical. Some of the
MBS implementation was stripped out as not necessary, so this code is closer to
the FUS version.

re: #5068
https://pulp.plan.io/issues/5068