Use relative_path instead of filename
The field filename seems to be redundant. Therefore, it's usages were removed from
the corresponding view set. Now, relative_path is used as filename instead.

In this commit, the field relative_path is no longer nullable. Also, the field
cannot be initialized to an empty string because it is used for a package creation
where the path is stored as location_href.

closes #4987
https://pulp.plan.io/issues/4987