Ensure updating erratum sets last_unit_added on *all* repos
When we update an erratum, this affects not only the repo
for which the update was requested, but also every other repo
containing the same erratum.

Therefore, rather than letting the controller only set
last_unit_added on the specific repo used for the import,
we should ensure the field is set for all repos containing
the unit - otherwise, publish of those repos may wrongly be
skipped.

closes #5951
https://pulp.plan.io/issues/5951