Publish a signed Yum repository if a signing service was attached
If a user creates a repository with a signing service, a publication will automatically contain a detached signature and a public key in addition. The detached signature and the public key are both used by a package manager in order to verify whether the repository is provided by a trusted authority or not.

closes #4812
https://pulp.plan.io/issues/4812