Refactor the method run()
In this commit, there were extracted multiple classes and methods from the method run() in order to improve the readability. The introduced classes RepositoryMetadataParser, ModulesMetadataParser, and PackagesComponentsParser are used for parsing the associated content within a repository. These classes alter data stored in the data class FirstStageData step by step, while parsing the content. In the end, this data class contains all necessary data required for downloading and processing the actual content.

[noissue]
[nocoverage]