Fix RPM copy w/ depsolving when packages depend on a particular file
Packages can depend on files e.g. /usr/bin/bash rather than "bash".
We're creating the file list metadata improperly in libsolv so this
wasn't being respected, and the problems would only be apparent in
debug log mode. During the period when we want to catch all these
bugs, "warning" would be more appropriate so that they are highly
visible.

closes: #7202
https://pulp.plan.io/issues/7202