Refactor to make metadata mirroring easier
So much of the logic is spread out that it makes it extremely difficult
to manage. e.g. all of the metadata is downloaded in many separate
functions. Also there are classes that would be better if they were just
functions, and helper methods that should really just be inlined because
they don't have a specific task.

closes: #8764
https://pulp.plan.io/issues/8764