Fix mirrored repositories for repos with distribution trees
Mirrored repository syncs were creating one publication per
sub-repository, which is in contrast to the way that pulp normally
works. Standard publish puts the sub-repos into the same publication
in a direct sub-directory because this is the only way to have
everything distributed properly. Otherwise the other publications are
in-accessible.

closes: #9098
https://pulp.plan.io/issues/9098