Add changelog limit to control the size of RPM metadata
Currently, no changelog limits are applied either when the
packages are uploaded or the metadata is published. This
resulted in extreme growth of other.xml metadata over time.
It's causing a multitude of problems like consuming excessive
memory and causing OOM while running "yum repoquery --changelog",
sync failure with timeout to download other.xml etc.
Hence, added a changelog limit, that will be picked from the
config or default to 10, to limit the number of changelogs to
be included in the xml.

Change-Id: I8595bff1a0832a7a79a9007b7b466416c9ed0849