Sort packages in metadata to improve compression [RHELDST-15791]
It was noticed the lack of ordering of packages in yum repo metadata
files reduced compression efficiency. These compressed files also
account for over half of the downloads. Improving the efficiency of
compression would thus reduce the overall cost of CDN.

This change sorts the unit_queryset. This will only apply to force /
full pushes, but investigation showed it would be challenging to make
this work with FastForward pushes.