Fixed drf JSONField being interpreted as Any
Drf JSONField is actually 'any' because it can be of any type (list, obj
int, bool, etc). Nevertheless, we (and drf-spectacular) always interpret
it as being of type 'object'.

In v0.27.0, drf-spectacular fixed that on their side, which broke us.

Although it is not correct to keep using JSONFields strictly as a object,
this commit adds a drf-spectacular extension that enforces that it should
always be trated as an 'object' on openapi schema generation.

The reason is to provide a safe fix for our users, who need their
bindings working the same way as they always did.

Closes: #3639