Handle chunked uploads when signing rpms
Currently chunked uploads are rejected when signing rpms with the following
message: "To sign a package on upload, a file must be provided."  This is
because the pulp cli sends the ID for the upload rather than a file.

To fix this, I've extended the serializer and viewset to handle both a file
and an upload, and have updated the signing task to properly deal with an
upload.

closes #3927

Signed-off-by: Jonathan Dieter <jdieter@ciq.com>