Always use "{NVRA}.rpm" for location_href and publications
Some remote repositories use arbitrary path+filenames for their packages,
which collides with how Pulp handles that. To avoid bugs caused by that,
we force the filename to be dervied from NVRA, which makes the publication
metadata and package listing unambigous.

On the sync task, Package now uses Package.filename as it's location_href.
On publication task, the Package.filename is used to the loation href on
the primary.xml metadata.

* Data Repair:
On systems that had synced 'bad' repositories before (and maybe didn't
realize it), the new changes will only fix the problem if this repository
and its packages are cleaned up from Pulp.

This add an rpm-datarepair command that makes it possible to fix the
sytem with a re-publish of the repository.

Closes: #4073
(cherry picked from commit 0067439a45b7ece39d658dea21c835aed2a69b46)