Ensure signing only happens when using `*_sign` policies
This commit adds two new sync policies, `additive_sign` and
`mirror_content_only_sign`, which are identical to the non `_sign`
versions, but will sign packages during the sync process.

It also adds a test to ensure that signing *doesn't* happen when not using
a `*_sign` policy.

And, because sync is smart enough to skip packages that have identical
nevras, we need to ensure that all packages are garbage collected before
running the tests, so we're adding the delete_orphans_pre fixture to both
tests.

Signed-off-by: Jonathan Dieter <jdieter@ciq.com>