During sync, skip package download for packages with unchanged NEVRAs
Currently, we start synchronizing packages even if the package is unchanged
from a previous sync.  This means that we end up re-signing packages each
time we sync when signing during sync.

This commit fixes that problem by re-using the already synced version of
the package for packages that already exist in the latest version of the
repository with the same NEVRA.  Note that this means that a package that
has been changed without changing the NEVRA will *not* be resynced anymore.
This may be seen as a security feature where packages can't be silently
changed.  To sync a package with the same NEVRA, the original package must
first be removed from the repository and then the repository must be synced
again.

Signed-off-by: Jonathan Dieter <jdieter@ciq.com>