Properly handle DeclarativeContent when checking for duplicate files
We need more details in the DeclarativeContent, so let's create it properly
and only use it there's at least one artifact.

Also, add a "already_in_repo" flag into `extra_data` to ensure that we
don't try to re-sign packages that were already in the repository (and
should already be signed).

Signed-off-by: Jonathan Dieter <jdieter@ciq.com>