Add test to ensure that artifact downloading in _sign_rpm_content works
Create a test that syncs a repository, then syncs a completely different
repository, and then re-syncs the first.  Artifacts will be created during
the first sync that will be orphaned (since signing creates a new
artifact), and then, on the third sync, packages will be deduped to these
artifacts.  This test verifies that the artifact downloading in
_sign_rpm_content works with these deduped artifacts.

Signed-off-by: Jonathan Dieter <jdieter@ciq.com>